fix(input system): altered how the inputs are implemented
