+++ MonstieWash/Assets/Scenes/MonstieWash.unity
@@ -493,6 +493,7 @@ GameObject:
   m_Component:
   - component: {fileID: 989945557}
   - component: {fileID: 989945556}
+  - component: {fileID: 989945558}
   - component: {fileID: 989945555}
   m_Layer: 0
   m_Name: hand_0
@@ -584,6 +585,20 @@ Transform:
   - {fileID: 26129519}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &989945558
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 989945554}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 07731d05c1d1d1943bddfe47bf8523ac, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_playerHand: {fileID: 989945555}
+  m_eraser: {fileID: 26129522}
 --- !u!1 &2064844343
 GameObject:
   m_ObjectHideFlags: 0
