+++ MonstieWash/Assets/Scripts/PlayerHandInput.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 07731d05c1d1d1943bddfe47bf8523ac
